abstract styles to styles.ts
